chore: move TestParticleCollection to test_data
